chore: code cleanup and organization
