+++ package.json
@@ -31,9 +31,9 @@
     "dev:starter-studio": "yarn --cwd dev/starter-studio dev",
     "dev:strict-studio": "yarn --cwd dev/strict-studio dev",
     "dev:test-studio": "yarn --cwd dev/test-studio dev",
-    "docs:report": "node -r dotenv-flow/config -r esbuild-register scripts/docReport",
-    "docs:report:create": "node -r dotenv-flow/config -r esbuild-register scripts/docReportCreate",
-    "docs:report:cleanup": "node -r dotenv-flow/config -r esbuild-register scripts/docReportCleanup",
+    "docs:report": "node -r dotenv-flow/config -r esbuild-register scripts/doc-report/docReport",
+    "docs:report:create": "node -r dotenv-flow/config -r esbuild-register scripts/doc-report/docReportCreate",
+    "docs:report:cleanup": "node -r dotenv-flow/config -r esbuild-register scripts/doc-report/docReportCleanup",
     "eslint": "eslint --ext=.cjs,.js,.jsx,.mjs,.ts,.tsx --quiet",
     "example:blog-studio": "yarn --cwd examples/blog-studio start",
     "example:clean-studio": "yarn --cwd examples/clean-studio start",
@@ -86,7 +86,7 @@
     "@babel/preset-env": "^7.19.1",
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.18.6",
-    "@bjoerge/mutiny": "^0.0.1",
+    "@bjoerge/mutiny": "^0.0.2",
     "@optimize-lodash/rollup-plugin": "^4.0.1",
     "@playwright/test": "^1.37.0",
     "@sanity/client": "^6.4.9",
