+++ scripts/doc-report/docClient.ts
@@ -1,0 +1,13 @@
+import {type SanityClient, createClient} from '@sanity/client'
+import {sanityIdify} from '../utils/sanityIdify'
+import {readEnv} from 'sanity-perf-tests/config/envVars'
+
+export function createDocClient(dataset: string): SanityClient {
+  return createClient({
+    projectId: 'c1zuxvqn',
+    dataset: sanityIdify(dataset),
+    token: readEnv('DOCS_REPORT_TOKEN'),
+    apiVersion: '2023-02-03',
+    useCdn: false,
+  })
+}
